Revert "[NUI] Make AccessibilityManager APIs as public ones (#1460)" (#1471)
This reverts commit 5056acf4fae2e2bbba0e1251c9fcb626a01ed039.
