+++ src/Tizen.NUI/src/internal/AccessibilityManager.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright(c) 2020 Samsung Electronics Co., Ltd.
+ * Copyright(c) 2019 Samsung Electronics Co., Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,23 +15,12 @@
  *
  */
 using System;
-using System.ComponentModel;
 using Tizen.NUI.BaseComponents;
 
 namespace Tizen.NUI
 {
-    /// <summary>
-    /// AccessibilityManager manages registration of views in an accessibility focus chain and changing the focused view within that chain.
-    /// This class provides the functionality of registering the focus order and description of views and maintaining the focus chain.
-    /// It provides functionality of setting the focus and moving the focus forward and backward.
-    ///  It also draws a highlight for the focused view and emits a signal when the focus is changed.
-    /// </summary>
-    /// <since_tizen> 6 </since_tizen>
-    /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-    [EditorBrowsable(EditorBrowsableState.Never)]
-    public partial class AccessibilityManager : BaseHandle
+    internal partial class AccessibilityManager : BaseHandle
     {
-        private static readonly AccessibilityManager instance = AccessibilityManager.Get();
 
         internal AccessibilityManager(global::System.IntPtr cPtr, bool cMemoryOwn) : base(Interop.AccessibilityManage.AccessibilityManager_SWIGUpcast(cPtr), cMemoryOwn)
         {
@@ -42,8 +31,6 @@ namespace Tizen.NUI
             return (obj == null) ? new global::System.Runtime.InteropServices.HandleRef(null, global::System.IntPtr.Zero) : obj.swigCPtr;
         }
 
-        /// This will not be public opened.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         protected override void ReleaseSwigCPtr(System.Runtime.InteropServices.HandleRef swigCPtr)
         {
             Interop.AccessibilityManage.delete_AccessibilityManager(swigCPtr);
@@ -52,10 +39,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager StatusChangedSignal
         private bool OnStatusChanged(IntPtr data)
         {
+            StatusChangedEventArgs e = new StatusChangedEventArgs();
+
+            // Populate all members of "e" (StatusChangedEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerStatusChangedEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerStatusChangedEventHandler(instance, null);
+                return _accessibilityManagerStatusChangedEventHandler(this, e);
             }
             return false;
         }
@@ -63,10 +55,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionNextSignal
         private bool OnActionNext(IntPtr data)
         {
+            ActionNextEventArgs e = new ActionNextEventArgs();
+
+            // Populate all members of "e" (ActionNextEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionNextEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionNextEventHandler(instance, null);
+                return _accessibilityManagerActionNextEventHandler(this, e);
             }
             return false;
         }
@@ -74,10 +71,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionPreviousSignal
         private bool OnActionPrevious(IntPtr data)
         {
+            ActionPreviousEventArgs e = new ActionPreviousEventArgs();
+
+            // Populate all members of "e" (ActionPreviousEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionPreviousEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionPreviousEventHandler(instance, null);
+                return _accessibilityManagerActionPreviousEventHandler(this, e);
             }
             return false;
         }
@@ -85,10 +87,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionActivateSignal
         private bool OnActionActivate(IntPtr data)
         {
+            ActionActivateEventArgs e = new ActionActivateEventArgs();
+
+            // Populate all members of "e" (ActionActivateEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionActivateEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionActivateEventHandler(instance, null);
+                return _accessibilityManagerActionActivateEventHandler(this, e);
             }
             return false;
         }
@@ -96,11 +103,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionReadSignal
         private bool OnActionRead(IntPtr data)
         {
+            ActionReadEventArgs e = new ActionReadEventArgs();
+
+            // Populate all members of "e" (ActionReadEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
 
             if (_accessibilityManagerActionReadEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionReadEventHandler(instance, null);
+                return _accessibilityManagerActionReadEventHandler(this, e);
             }
             return false;
         }
@@ -108,10 +119,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionOverSignal
         private bool OnActionOver(IntPtr data)
         {
+            ActionOverEventArgs e = new ActionOverEventArgs();
+
+            // Populate all members of "e" (ActionOverEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionOverEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionOverEventHandler(instance, null);
+                return _accessibilityManagerActionOverEventHandler(this, e);
             }
             return false;
         }
@@ -119,21 +135,32 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionReadNextSignal
         private bool OnActionReadNext(IntPtr data)
         {
+            ActionReadNextEventArgs e = new ActionReadNextEventArgs();
+
+            // Populate all members of "e" (ActionReadNextEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionReadNextEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionReadNextEventHandler(instance, null);
+                return _accessibilityManagerActionReadNextEventHandler(this, e);
             }
             return false;
         }
 
+
         // Callback for AccessibilityManager ActionReadPreviousSignal
         private bool OnActionReadPrevious(IntPtr data)
         {
+            ActionReadPreviousEventArgs e = new ActionReadPreviousEventArgs();
+
+            // Populate all members of "e" (ActionReadPreviousEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionReadPreviousEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionReadPreviousEventHandler(instance, null);
+                return _accessibilityManagerActionReadPreviousEventHandler(this, e);
             }
             return false;
         }
@@ -141,10 +168,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionUpSignal
         private bool OnActionUp(IntPtr data)
         {
+            ActionUpEventArgs e = new ActionUpEventArgs();
+
+            // Populate all members of "e" (ActionUpEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionUpEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionUpEventHandler(instance, null);
+                return _accessibilityManagerActionUpEventHandler(this, e);
             }
             return false;
         }
@@ -152,10 +184,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionDownSignal
         private bool OnActionDown(IntPtr data)
         {
+            ActionDownEventArgs e = new ActionDownEventArgs();
+
+            // Populate all members of "e" (ActionDownEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionDownEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionDownEventHandler(instance, null);
+                return _accessibilityManagerActionDownEventHandler(this, e);
             }
             return false;
         }
@@ -163,10 +200,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionClearFocusSignal
         private bool OnActionClearFocus(IntPtr data)
         {
+            ActionClearFocusEventArgs e = new ActionClearFocusEventArgs();
+
+            // Populate all members of "e" (ActionClearFocusEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionClearFocusEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionClearFocusEventHandler(instance, null);
+                return _accessibilityManagerActionClearFocusEventHandler(this, e);
             }
             return false;
         }
@@ -174,10 +216,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionBackSignal
         private bool OnActionBack(IntPtr data)
         {
+            ActionBackEventArgs e = new ActionBackEventArgs();
+
+            // Populate all members of "e" (ActionBackEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionBackEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionBackEventHandler(instance, null);
+                return _accessibilityManagerActionBackEventHandler(this, e);
             }
             return false;
         }
@@ -185,10 +232,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionScrollUpSignal
         private bool OnActionScrollUp(IntPtr data)
         {
+            ActionScrollUpEventArgs e = new ActionScrollUpEventArgs();
+
+            // Populate all members of "e" (ActionScrollUpEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionScrollUpEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionScrollUpEventHandler(instance, null);
+                return _accessibilityManagerActionScrollUpEventHandler(this, e);
             }
             return false;
         }
@@ -196,21 +248,32 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionScrollDownSignal
         private bool OnActionScrollDown(IntPtr data)
         {
+            ActionScrollDownEventArgs e = new ActionScrollDownEventArgs();
+
+            // Populate all members of "e" (ActionScrollDownEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionScrollDownEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionScrollDownEventHandler(instance, null);
+                return _accessibilityManagerActionScrollDownEventHandler(this, e);
             }
             return false;
         }
 
+
         // Callback for AccessibilityManager ActionPageLeftSignal
         private bool OnActionPageLeft(IntPtr data)
         {
+            ActionPageLeftEventArgs e = new ActionPageLeftEventArgs();
+
+            // Populate all members of "e" (ActionPageLeftEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionPageLeftEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionPageLeftEventHandler(instance, null);
+                return _accessibilityManagerActionPageLeftEventHandler(this, e);
             }
             return false;
         }
@@ -218,10 +281,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionPageRightSignal
         private bool OnActionPageRight(IntPtr data)
         {
+            ActionPageRightEventArgs e = new ActionPageRightEventArgs();
+
+            // Populate all members of "e" (ActionPageRightEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionPageRightEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionPageRightEventHandler(instance, null);
+                return _accessibilityManagerActionPageRightEventHandler(this, e);
             }
             return false;
         }
@@ -229,10 +297,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionPageUpSignal
         private bool OnActionPageUp(IntPtr data)
         {
+            ActionPageUpEventArgs e = new ActionPageUpEventArgs();
+
+            // Populate all members of "e" (ActionPageUpEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionPageUpEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionPageUpEventHandler(instance, null);
+                return _accessibilityManagerActionPageUpEventHandler(this, e);
             }
             return false;
         }
@@ -241,10 +314,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionPageDownSignal
         private bool OnActionPageDown(IntPtr data)
         {
+            ActionPageDownEventArgs e = new ActionPageDownEventArgs();
+
+            // Populate all members of "e" (ActionPageDownEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionPageDownEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionPageDownEventHandler(instance, null);
+                return _accessibilityManagerActionPageDownEventHandler(this, e);
             }
             return false;
         }
@@ -253,10 +331,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionMoveToFirstSignal
         private bool OnActionMoveToFirst(IntPtr data)
         {
+            ActionMoveToFirstEventArgs e = new ActionMoveToFirstEventArgs();
+
+            // Populate all members of "e" (ActionMoveToFirstEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionMoveToFirstEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionMoveToFirstEventHandler(instance, null);
+                return _accessibilityManagerActionMoveToFirstEventHandler(this, e);
             }
             return false;
         }
@@ -264,10 +347,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionMoveToLastSignal
         private bool OnActionMoveToLast(IntPtr data)
         {
+            ActionMoveToLastEventArgs e = new ActionMoveToLastEventArgs();
+
+            // Populate all members of "e" (ActionMoveToLastEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionMoveToLastEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionMoveToLastEventHandler(instance, null);
+                return _accessibilityManagerActionMoveToLastEventHandler(this, e);
             }
             return false;
         }
@@ -275,10 +363,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionReadFromTopSignal
         private bool OnActionReadFromTop(IntPtr data)
         {
+            ActionReadFromTopEventArgs e = new ActionReadFromTopEventArgs();
+
+            // Populate all members of "e" (ActionReadFromTopEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionReadFromTopEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionReadFromTopEventHandler(instance, null);
+                return _accessibilityManagerActionReadFromTopEventHandler(this, e);
             }
             return false;
         }
@@ -286,10 +379,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionReadFromNextSignal
         private bool OnActionReadFromNext(IntPtr data)
         {
+            ActionReadFromNextEventArgs e = new ActionReadFromNextEventArgs();
+
+            // Populate all members of "e" (ActionReadFromNextEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionReadFromNextEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionReadFromNextEventHandler(instance, null);
+                return _accessibilityManagerActionReadFromNextEventHandler(this, e);
             }
             return false;
         }
@@ -297,10 +395,31 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionZoomSignal
         private bool OnActionZoom(IntPtr data)
         {
+            ActionZoomEventArgs e = new ActionZoomEventArgs();
+
+            // Populate all members of "e" (ActionZoomEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionZoomEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionZoomEventHandler(instance, null);
+                return _accessibilityManagerActionZoomEventHandler(this, e);
+            }
+            return false;
+        }
+
+        // Callback for AccessibilityManager ActionReadIndicatorInformationSignal
+        private bool OnActionReadIndicatorInformation(IntPtr data)
+        {
+            ActionReadIndicatorInformationEventArgs e = new ActionReadIndicatorInformationEventArgs();
+
+            // Populate all members of "e" (ActionReadIndicatorInformationEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
+            if (_accessibilityManagerActionReadIndicatorInformationEventHandler != null)
+            {
+                //here we send all data to user event handlers
+                return _accessibilityManagerActionReadIndicatorInformationEventHandler(this, e);
             }
             return false;
         }
@@ -308,10 +427,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionReadPauseResumeSignal
         private bool OnActionReadPauseResume(IntPtr data)
         {
+            ActionReadPauseResumeEventArgs e = new ActionReadPauseResumeEventArgs();
+
+            // Populate all members of "e" (ActionReadPauseResumeEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionReadPauseResumeEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionReadPauseResumeEventHandler(instance, null);
+                return _accessibilityManagerActionReadPauseResumeEventHandler(this, e);
             }
             return false;
         }
@@ -319,10 +443,15 @@ namespace Tizen.NUI
         // Callback for AccessibilityManager ActionStartStopSignal
         private bool OnActionStartStop(IntPtr data)
         {
+            ActionStartStopEventArgs e = new ActionStartStopEventArgs();
+
+            // Populate all members of "e" (ActionStartStopEventArgs) with real data
+            e.AccessibilityManager = AccessibilityManager.GetAccessibilityManagerFromPtr(data);
+
             if (_accessibilityManagerActionStartStopEventHandler != null)
             {
                 //here we send all data to user event handlers
-                return _accessibilityManagerActionStartStopEventHandler(instance, null);
+                return _accessibilityManagerActionStartStopEventHandler(this, e);
             }
             return false;
         }
@@ -374,128 +503,33 @@ namespace Tizen.NUI
             }
         }
 
-        /// <summary>
-        /// Enumeration for accessibility that needs four information which will be read by screen-reader.
-        ///
-        /// Reading order : Label -> Trait -> Optional (Value and Hint)
-        /// </summary>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public enum AccessibilityAttribute
-        {
-            /// <summary>
-            /// Simple text which contained in components, such as Ok or Cancel in case of a button
-            /// </summary>
-            /// <since_tizen> 6 </since_tizen>
-            /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-            [EditorBrowsable(EditorBrowsableState.Never)]
-            Label = 0,
-            /// <summary>
-            /// Description of components trait, such as Button in case of a button
-            /// </summary>
-            /// <since_tizen> 6 </since_tizen>
-            /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-            [EditorBrowsable(EditorBrowsableState.Never)]
-            Trait,
-            /// <summary>
-            /// Current value of components (Optional)
-            /// </summary>
-            /// <since_tizen> 6 </since_tizen>
-            /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-            [EditorBrowsable(EditorBrowsableState.Never)]
-            Value,
-            /// <summary>
-            /// Hint for action (Optional)
-            /// </summary>
-            /// <since_tizen> 6 </since_tizen>
-            /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-            [EditorBrowsable(EditorBrowsableState.Never)]
-            Hint,
-            /// <summary>
-            /// The number of attributes
-            /// </summary>
-            /// <since_tizen> 6 </since_tizen>
-            /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-            [EditorBrowsable(EditorBrowsableState.Never)]
-            AttributeNumber
-        }
 
-        /// <summary>
-        /// Enumeration for overshoot direction.
-        /// </summary>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public enum FocusOvershotDirection
+        public static AccessibilityManager GetAccessibilityManagerFromPtr(global::System.IntPtr cPtr)
         {
-            /// <summary>
-            /// Try to move previous of the first view
-            /// </summary>
-            /// <since_tizen> 6 </since_tizen>
-            /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-            [EditorBrowsable(EditorBrowsableState.Never)]
-            Previous = -1,
-            /// <summary>
-            /// Try to move next of the last view
-            /// </summary>
-            /// <since_tizen> 6 </since_tizen>
-            /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-            [EditorBrowsable(EditorBrowsableState.Never)]
-            Next = 1
+            AccessibilityManager ret = new AccessibilityManager(cPtr, false);
+            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
+            return ret;
         }
 
 
-        /// <summary>
-        /// Creates an AccessibilityManager handle.
-        /// </summary>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public AccessibilityManager() : this(Interop.AccessibilityManage.new_AccessibilityManager(), true)
         {
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        /// <summary>
-        /// Gets the singleton of AccessibilityManager object.
-        /// </summary>
-        /// <returns> A handle to the AccessibilityManager </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public static AccessibilityManager Instance
+        public static AccessibilityManager Get()
         {
-            get
-            {
-                return instance;
-            }
+            AccessibilityManager ret = new AccessibilityManager(Interop.AccessibilityManage.AccessibilityManager_Get(), true);
+            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
+            return ret;
         }
 
-        /// <summary>
-        /// Sets the information of the specified view's accessibility attribute.
-        /// </summary>
-        /// <param name="view"> The view to be set with</param>
-        /// <param name="type"> The attribute type the text to be set with</param>
-        /// <param name="text"> The text for the view's accessibility information</param>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public void SetAccessibilityAttribute(View view, AccessibilityManager.AccessibilityAttribute type, string text)
         {
             Interop.AccessibilityManage.AccessibilityManager_SetAccessibilityAttribute(swigCPtr, View.getCPtr(view), (int)type, text);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        /// <summary>
-        /// Gets the text of the specified view's accessibility attribute.
-        /// </summary>
-        /// <param name="view"> The view to be queried</param>
-        /// <param name="type"> The attribute type to be queried</param>
-        /// <returns> The text of the view's accessibility information </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public string GetAccessibilityAttribute(View view, AccessibilityManager.AccessibilityAttribute type)
         {
             string ret = Interop.AccessibilityManage.AccessibilityManager_GetAccessibilityAttribute(swigCPtr, View.getCPtr(view), (int)type);
@@ -503,34 +537,12 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Sets the focus order of the view.
-        /// The focus order of each view in the focus chain is unique.
-        /// If there is another view assigned with the same focus order already, the new view will be inserted to the focus chain with that focus order,
-        /// and the focus order of the original view and all the views followed in the focus chain will be increased accordingly.
-        /// If the focus order assigned to the view is 0, it means that view's focus order is undefined
-        /// (e.g. the view has a  description but with no focus order being set yet) and therefore that view is not focusable.
-        /// </summary>
-        /// <param name="view"> the view to be set with</param>
-        /// <param name="order"> the focus order to be set with</param>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public void SetFocusOrder(View view, uint order)
         {
             Interop.AccessibilityManage.AccessibilityManager_SetFocusOrder(swigCPtr, View.getCPtr(view), order);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        /// <summary>
-        /// Gets the focus order of the view.
-        /// When the focus order is 0, it means the focus order of the view is undefined.
-        /// </summary>
-        /// <param name="view"> the view to be set with</param>
-        /// <returns> The focus order of the view </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public uint GetFocusOrder(View view)
         {
             uint ret = Interop.AccessibilityManage.AccessibilityManager_GetFocusOrder(swigCPtr, View.getCPtr(view));
@@ -538,17 +550,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Generates a new focus order number which can be used to assign to views
-        /// which need to be appended to the end of the current focus order chain.
-        /// The new number will be an increment over the very last focus order number in the focus chain.
-        /// If the focus chain is empty then the function returns 1,
-        /// else the number returned will be FOLast + 1 where FOLast is the focus order of the very last control in the focus chain.
-        /// </summary>
-        /// <returns> The focus order of the view </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public uint GenerateNewFocusOrder()
         {
             uint ret = Interop.AccessibilityManage.AccessibilityManager_GenerateNewFocusOrder(swigCPtr);
@@ -556,15 +557,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Gets the view that has the specified focus order.
-        /// It will return an empty handle if no view in the window has the specified focus order.
-        /// </summary>
-        /// <param name="order"> The focus order of the view</param>
-        /// <returns> The view that has the specified focus order or an empty handle if no view in the stage has the specified focus order </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public View GetViewByFocusOrder(uint order)
         {
             View ret = new View(Interop.AccessibilityManage.AccessibilityManager_GetActorByFocusOrder(swigCPtr, order), true);
@@ -572,16 +564,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Moves the focus to the specified view.
-        /// Only one view can be focused at the same time. The view must have a defined focus order
-        /// and must be focusable, visible and in the window.
-        /// </summary>
-        /// <param name="view"> the view to be set with</param>
-        /// <returns> Whether the focus is successful or not </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public bool SetCurrentFocusView(View view)
         {
             bool ret = Interop.AccessibilityManage.AccessibilityManager_SetCurrentFocusActor(swigCPtr, View.getCPtr(view));
@@ -589,13 +571,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Gets the current focused view.
-        /// </summary>
-        /// <returns> A handle to the current focused view or an empty handle if no view is focused </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public View GetCurrentFocusView()
         {
             View ret = new View(Interop.AccessibilityManage.AccessibilityManager_GetCurrentFocusActor(swigCPtr), true);
@@ -603,13 +578,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Gets the focus group of current focused view.
-        /// </summary>
-        /// <returns> A handle to the immediate parent of the current focused view which is also a focus group, or an empty handle if no view is focused </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public View GetCurrentFocusGroup()
         {
             View ret = new View(Interop.AccessibilityManage.AccessibilityManager_GetCurrentFocusGroup(swigCPtr), true);
@@ -617,13 +585,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Gets the focus order of currently focused view.
-        /// </summary>
-        /// <returns> The focus order of the currently focused view or 0 if no view is in focus </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public uint GetCurrentFocusOrder()
         {
             uint ret = Interop.AccessibilityManage.AccessibilityManager_GetCurrentFocusOrder(swigCPtr);
@@ -631,14 +592,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Moves the focus to the next focusable view in the focus chain (according to the focus traversal order).
-        /// When the focus movement is wrapped around, the focus will be moved to the first focusable view when it reaches the end of the focus chain.
-        /// </summary>
-        /// <returns> True if the moving was successful </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public bool MoveFocusForward()
         {
             bool ret = Interop.AccessibilityManage.AccessibilityManager_MoveFocusForward(swigCPtr);
@@ -646,15 +599,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Moves the focus to the previous focusable view in the focus chain (according to the focus traversal order).
-        /// When the focus movement is wrapped around, the focus will be moved to the last focusable view
-        /// when it reaches the beginning of the focus chain.
-        /// </summary>
-        /// <returns> True if the moving was successful </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public bool MoveFocusBackward()
         {
             bool ret = Interop.AccessibilityManage.AccessibilityManager_MoveFocusBackward(swigCPtr);
@@ -662,53 +606,24 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Clears the focus from the current focused view if any, so that no view is focused in the focus chain.
-        /// It will emit focus changed signal without current focused view.
-        /// </summary>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public void ClearFocus()
         {
             Interop.AccessibilityManage.AccessibilityManager_ClearFocus(swigCPtr);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        /// <summary>
-        /// Clears every registered focusable view from focus-manager.
-        /// </summary>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public new void Reset()
         {
             Interop.AccessibilityManage.AccessibilityManager_Reset(swigCPtr);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        /// <summary>
-        /// Sets whether an view is a focus group that can limit the scope of focus movement to its child views in the focus chain.
-        /// </summary>
-        /// <param name="view"> the view to be set as a focus group</param>
-        /// <param name="isFocusGroup"> Whether to set the view to be a focus group or not</param>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public void SetFocusGroup(View view, bool isFocusGroup)
         {
             Interop.AccessibilityManage.AccessibilityManager_SetFocusGroup(swigCPtr, View.getCPtr(view), isFocusGroup);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        /// <summary>
-        /// Checks whether the view is set as a focus group or not.
-        /// </summary>
-        /// <param name="view"> the view to be checked</param>
-        /// <returns> Whether the view is set as a focus group </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public bool IsFocusGroup(View view)
         {
             bool ret = Interop.AccessibilityManage.AccessibilityManager_IsFocusGroup(swigCPtr, View.getCPtr(view));
@@ -716,28 +631,12 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Sets whether the group mode is enabled or not.
-        /// When the group mode is enabled, the focus movement will be limited to the child views  of the current focus group including the current focus group itself.
-        /// The current focus group is the closest ancestor of the current focused view that is set as a focus group.
-        /// </summary>
-        /// <param name="enabled"> Whether the group mode is enabled or not</param>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public void SetGroupMode(bool enabled)
         {
             Interop.AccessibilityManage.AccessibilityManager_SetGroupMode(swigCPtr, enabled);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        /// <summary>
-        /// Gets whether the group mode is enabled or not.
-        /// </summary>
-        /// <since_tizen> 6 </since_tizen>
-        /// <returns> Whether the group mode is enabled or not. </returns>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public bool GetGroupMode()
         {
             bool ret = Interop.AccessibilityManage.AccessibilityManager_GetGroupMode(swigCPtr);
@@ -745,28 +644,12 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Sets whether focus will be moved to the beginning of the focus chain when it reaches the end or vice versa.
-        /// When both the wrap mode and the group mode are enabled, focus will be wrapped within the current focus group.
-        /// Focus will not be wrapped in default.
-        /// </summary>
-        /// <param name="wrapped"> Whether the focus movement is wrapped around or not</param>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public void SetWrapMode(bool wrapped)
         {
             Interop.AccessibilityManage.AccessibilityManager_SetWrapMode(swigCPtr, wrapped);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        /// <summary>
-        /// Gets whether the wrap mode is enabled or not.
-        /// </summary>
-        /// <returns> Whether the wrap mode is enabled or not. </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public bool GetWrapMode()
         {
             bool ret = Interop.AccessibilityManage.AccessibilityManager_GetWrapMode(swigCPtr);
@@ -774,28 +657,12 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Sets the focus indicator view.
-        /// This will replace the default focus indicator view in AccessibilityManager and
-        /// will be added to the focused view as a highlight.
-        /// </summary>
-        /// <param name="indicator"> The indicator view to be added</param>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public void SetFocusIndicatorView(View indicator)
         {
             Interop.AccessibilityManage.AccessibilityManager_SetFocusIndicatorActor(swigCPtr, View.getCPtr(indicator));
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
         }
 
-        /// <summary>
-        /// Gets the focus indicator view.
-        /// </summary>
-        /// <returns> A handle to the focus indicator view </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public View GetFocusIndicatorView()
         {
             View ret = new View(Interop.AccessibilityManage.AccessibilityManager_GetFocusIndicatorActor(swigCPtr), true);
@@ -803,14 +670,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Returns the closest ancestor of the given view that is a focus group.
-        /// </summary>
-        /// <param name="view"> The view to be checked for its focus group</param>
-        /// <returns> The focus group the given view belongs to or an empty handle if the given view doesn't belong to any focus group </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public View GetFocusGroup(View view)
         {
             View ret = new View(Interop.AccessibilityManage.AccessibilityManager_GetFocusGroup(swigCPtr, View.getCPtr(view)), true);
@@ -818,13 +677,6 @@ namespace Tizen.NUI
             return ret;
         }
 
-        /// <summary>
-        /// Returns the current position of the read action.
-        /// </summary>
-        /// <returns> The current event position </returns>
-        /// <since_tizen> 6 </since_tizen>
-        /// This will be public opened in tizen_6.0 after ACR done. Before ACR, need to be hidden as inhouse API.
-        [EditorBrowsable(EditorBrowsableState.Never)]
         public Vector2 GetReadPosition()
         {
             Vector2 ret = new Vector2(Interop.AccessibilityManage.AccessibilityManager_GetReadPosition(swigCPtr), true);
@@ -832,215 +684,231 @@ namespace Tizen.NUI
             return ret;
         }
 
-        internal static AccessibilityManager Get()
-        {
-            AccessibilityManager ret = new AccessibilityManager(Interop.AccessibilityManage.AccessibilityManager_Get(), true);
-            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
-            return ret;
-        }
-
-        // Signals - AccessibilityManagerEvent.cs
-        internal FocusChangedSignal FocusChangedSignal()
+        public FocusChangedSignal FocusChangedSignal()
         {
             FocusChangedSignal ret = new FocusChangedSignal(Interop.AccessibilityManage.AccessibilityManager_FocusChangedSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityFocusOvershotSignal FocusOvershotSignal()
+        public AccessibilityFocusOvershotSignal FocusOvershotSignal()
         {
             AccessibilityFocusOvershotSignal ret = new AccessibilityFocusOvershotSignal(Interop.AccessibilityManage.AccessibilityManager_FocusOvershotSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal ViewSignal FocusedViewActivatedSignal()
+        public ViewSignal FocusedViewActivatedSignal()
         {
             ViewSignal ret = new ViewSignal(Interop.AccessibilityManage.AccessibilityManager_FocusedActorActivatedSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal StatusChangedSignal()
+        public AccessibilityActionSignal StatusChangedSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_StatusChangedSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionNextSignal()
+        public AccessibilityActionSignal ActionNextSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionNextSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionPreviousSignal()
+        public AccessibilityActionSignal ActionPreviousSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionPreviousSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionActivateSignal()
+        public AccessibilityActionSignal ActionActivateSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionActivateSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionReadSignal()
+        public AccessibilityActionSignal ActionReadSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionReadSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionOverSignal()
+        public AccessibilityActionSignal ActionOverSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionOverSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionReadNextSignal()
+        public AccessibilityActionSignal ActionReadNextSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionReadNextSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionReadPreviousSignal()
+        public AccessibilityActionSignal ActionReadPreviousSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionReadPreviousSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionUpSignal()
+        public AccessibilityActionSignal ActionUpSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionUpSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionDownSignal()
+        public AccessibilityActionSignal ActionDownSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionDownSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionClearFocusSignal()
+        public AccessibilityActionSignal ActionClearFocusSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionClearFocusSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionBackSignal()
+        public AccessibilityActionSignal ActionBackSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionBackSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionScrollUpSignal()
+        public AccessibilityActionSignal ActionScrollUpSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionScrollUpSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionScrollDownSignal()
+        public AccessibilityActionSignal ActionScrollDownSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionScrollDownSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionPageLeftSignal()
+        public AccessibilityActionSignal ActionPageLeftSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionPageLeftSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionPageRightSignal()
+        public AccessibilityActionSignal ActionPageRightSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionPageRightSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionPageUpSignal()
+        public AccessibilityActionSignal ActionPageUpSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionPageUpSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionPageDownSignal()
+        public AccessibilityActionSignal ActionPageDownSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionPageDownSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionMoveToFirstSignal()
+        public AccessibilityActionSignal ActionMoveToFirstSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionMoveToFirstSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionMoveToLastSignal()
+        public AccessibilityActionSignal ActionMoveToLastSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionMoveToLastSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionReadFromTopSignal()
+        public AccessibilityActionSignal ActionReadFromTopSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionReadFromTopSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionReadFromNextSignal()
+        public AccessibilityActionSignal ActionReadFromNextSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionReadFromNextSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionZoomSignal()
+        public AccessibilityActionSignal ActionZoomSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionZoomSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionReadPauseResumeSignal()
+        public AccessibilityActionSignal ActionReadIndicatorInformationSignal()
+        {
+            AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionReadIndicatorInformationSignal(swigCPtr), false);
+            if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
+            return ret;
+        }
+
+        public AccessibilityActionSignal ActionReadPauseResumeSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionReadPauseResumeSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal AccessibilityActionSignal ActionStartStopSignal()
+        public AccessibilityActionSignal ActionStartStopSignal()
         {
             AccessibilityActionSignal ret = new AccessibilityActionSignal(Interop.AccessibilityManage.AccessibilityManager_ActionStartStopSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
         }
 
-        internal SWIGTYPE_p_Dali__SignalT_bool_fDali__Toolkit__AccessibilityManager_R_Dali__TouchEvent_const_RF_t ActionScrollSignal()
+        public SWIGTYPE_p_Dali__SignalT_bool_fDali__Toolkit__AccessibilityManager_R_Dali__TouchEvent_const_RF_t ActionScrollSignal()
         {
             SWIGTYPE_p_Dali__SignalT_bool_fDali__Toolkit__AccessibilityManager_R_Dali__TouchEvent_const_RF_t ret = new SWIGTYPE_p_Dali__SignalT_bool_fDali__Toolkit__AccessibilityManager_R_Dali__TouchEvent_const_RF_t(Interop.AccessibilityManage.AccessibilityManager_ActionScrollSignal(swigCPtr), false);
             if (NDalicPINVOKE.SWIGPendingException.Pending) throw NDalicPINVOKE.SWIGPendingException.Retrieve();
             return ret;
+        }
+
+        /// <since_tizen> 3 </since_tizen>
+        public enum AccessibilityAttribute
+        {
+            ACCESSIBILITY_LABEL = 0,
+            ACCESSIBILITY_TRAIT,
+            ACCESSIBILITY_VALUE,
+            ACCESSIBILITY_HINT,
+            ACCESSIBILITY_ATTRIBUTE_NUM
+        }
+
+        /// <since_tizen> 3 </since_tizen>
+        public enum FocusOvershotDirection
+        {
+            OVERSHOT_PREVIOUS = -1,
+            OVERSHOT_NEXT = 1
         }
     }
 }
